mount_point example
