+++ examples/two_apps/README.md
@@ -10,3 +10,6 @@ The example illustrates how one can take control of the mounting process which i
 
 Instead of using the same component type twice, the example could use two entirely different components that communicate with each other.
 One of the components could even accept a generic "remote" component using a trait.
+
+This example is very similar to [`mount_point`](../mount_point).
+The two should be merged into a single example.
